+++ src/uwtools/drivers/fv3.py
@@ -11,7 +11,7 @@ from iotaa import asset, task, tasks
 
 from uwtools.config.formats.nml import NMLConfig
 from uwtools.config.formats.yaml import YAMLConfig
-from uwtools.drivers.standalonedriver import StandaloneDriver
+from uwtools.drivers.driver import StandaloneDriver
 from uwtools.logging import log
 from uwtools.strings import STR
 from uwtools.utils.tasks import filecopy, symlink
