+++ src/uwtools/drivers/mpas.py
@@ -10,7 +10,7 @@ from iotaa import asset, task, tasks
 
 from uwtools.api.template import render
 from uwtools.config.formats.nml import NMLConfig
-from uwtools.drivers.standalonedriver import StandaloneDriver
+from uwtools.drivers.driver import StandaloneDriver
 from uwtools.exceptions import UWConfigError
 from uwtools.strings import STR
 from uwtools.utils.tasks import file, filecopy, symlink
