+++ src/uwtools/drivers/sfc_climo_gen.py
@@ -8,7 +8,7 @@ from typing import List, Optional
 from iotaa import asset, task, tasks
 
 from uwtools.config.formats.nml import NMLConfig
-from uwtools.drivers.standalonedriver import StandaloneDriver
+from uwtools.drivers.driver import StandaloneDriver
 from uwtools.strings import STR
 from uwtools.utils.tasks import file
