--- src/uwtools/drivers/standalonedriver.py
@@ -1,186 +1,0 @@
-"""
-An abstract class for component drivers.
-"""
-
-import os
-import re
-import stat
-from datetime import datetime, timedelta
-from pathlib import Path
-from textwrap import dedent
-from typing import Any, Dict, List, Optional, Union
-
-from iotaa import asset, task, tasks
-
-from uwtools.drivers.driver import Driver
-from uwtools.exceptions import UWConfigError
-from uwtools.scheduler import JobScheduler
-from uwtools.utils.processing import execute
-
-
-class StandaloneDriver(Driver):
-    """
-    An abstract class for component drivers.
-    """
-
-    def __init__(
-        self,
-        config: Optional[Union[dict, Path]],
-        dry_run: bool = False,
-        batch: bool = False,
-        cycle: Optional[datetime] = None,
-        key_path: Optional[List[str]] = None,
-        leadtime: Optional[timedelta] = None,
-    ) -> None:
-        """
-        A component driver.
-
-        :param config: Path to config file (read stdin if missing or None).
-        :param dry_run: Run in dry-run mode?
-        :param batch: Run component via the batch system?
-        :param cycle: The cycle.
-        :param key_path: Keys leading through the config to the driver's configuration block.
-        :param leadtime: The leadtime.
-        """
-        super().__init__(
-            config=config,
-            dry_run=dry_run,
-            batch=batch,
-            cycle=cycle,
-            key_path=key_path,
-            leadtime=leadtime,
-        )
-
-    # Workflow tasks
-
-    @tasks
-    def run(self):
-        """
-        A run.
-        """
-        yield self._taskname("run")
-        yield (self._run_via_batch_submission() if self._batch else self._run_via_local_execution())
-
-    @task
-    def _run_via_batch_submission(self):
-        """
-        A run executed via the batch system.
-        """
-        yield self._taskname("run via batch submission")
-        path = Path("%s.submit" % self._runscript_path)
-        yield asset(path, path.is_file)
-        yield self.provisioned_run_directory()
-        self._scheduler.submit_job(runscript=self._runscript_path, submit_file=path)
-
-    @task
-    def _run_via_local_execution(self):
-        """
-        A run executed directly on the local system.
-        """
-        yield self._taskname("run via local execution")
-        path = self._rundir / f"done.{self._driver_name}"
-        yield asset(path, path.is_file)
-        yield self.provisioned_run_directory()
-        cmd = "{x} >{x}.out 2>&1".format(x=self._runscript_path)
-        execute(cmd=cmd, cwd=self._rundir, log_output=True)
-
-    # Private helper methods
-
-    @property
-    def _resources(self) -> Dict[str, Any]:
-        """
-        Returns configuration data for the runscript.
-        """
-        try:
-            platform = self._config["platform"]
-        except KeyError as e:
-            raise UWConfigError("Required 'platform' block missing in config") from e
-        return {
-            "account": platform["account"],
-            "rundir": self._rundir,
-            "scheduler": platform["scheduler"],
-            "stdout": "%s.out" % self._runscript_path.name,  # config may override
-            **self._driver_config.get("execution", {}).get("batchargs", {}),
-        }
-
-    @property
-    def _runcmd(self) -> str:
-        """
-        Returns the full command-line component invocation.
-        """
-        execution = self._driver_config.get("execution", {})
-        mpiargs = execution.get("mpiargs", [])
-        components = [
-            execution.get("mpicmd"),  # MPI run program
-            *[str(x) for x in mpiargs],  # MPI arguments
-            execution["executable"],  # component executable name
-        ]
-        return " ".join(filter(None, components))
-
-    def _runscript(
-        self,
-        execution: List[str],
-        envcmds: Optional[List[str]] = None,
-        envvars: Optional[Dict[str, str]] = None,
-        scheduler: Optional[JobScheduler] = None,
-    ) -> str:
-        """
-        Returns a driver runscript.
-
-        :param execution: Statements to execute.
-        :param envcmds: Shell commands to set up runtime environment.
-        :param envvars: Environment variables to set in runtime environment.
-        :param scheduler: A job-scheduler object.
-        """
-        # Render script sections into a template, remove any extraneous newlines related to elided
-        # sections, then return the resulting string.
-        template = """
-        #!/bin/bash
-
-        {directives}
-
-        {envcmds}
-
-        {envvars}
-
-        {execution}
-        """
-        rs = dedent(template).format(
-            directives="\n".join(scheduler.directives if scheduler else ""),
-            envcmds="\n".join(envcmds or []),
-            envvars="\n".join([f"export {k}={v}" for k, v in (envvars or {}).items()]),
-            execution="\n".join(execution),
-        )
-        return re.sub(r"\n\n\n+", "\n\n", rs.strip())
-
-    @property
-    def _runscript_path(self) -> Path:
-        """
-        Returns the path to the runscript.
-        """
-        return self._rundir / f"runscript.{self._driver_name}"
-
-    @property
-    def _scheduler(self) -> JobScheduler:
-        """
-        Returns the job scheduler specified by the platform information.
-        """
-        return JobScheduler.get_scheduler(self._resources)
-
-    def _write_runscript(self, path: Path, envvars: Dict[str, str]) -> None:
-        """
-        Write the runscript.
-        """
-        path.parent.mkdir(parents=True, exist_ok=True)
-        rs = self._runscript(
-            envcmds=self._driver_config.get("execution", {}).get("envcmds", []),
-            envvars=envvars,
-            execution=[
-                "time %s" % self._runcmd,
-                "test $? -eq 0 && touch %s.done" % self._runscript_path.name,
-            ],
-            scheduler=self._scheduler if self._batch else None,
-        )
-        with open(path, "w", encoding="utf-8") as f:
-            print(rs, file=f)
-        os.chmod(path, os.stat(path).st_mode | stat.S_IEXEC)
